Correct an issue with generation of column labels
An issue with the generation of columns labels while loading data from a
file resulted in a None value or empty list in a case where a valid list
of column labels should have been generated. Correct this issue and also
update hello_docket script to print a summary of file metadata.