Complete hello_docket example
For deploying on development machines, a basic example script is needed
to ensure that dependencies are present. If the example script runs and
prints expected output, this indicates the development machine has the
necessary dependencies. hello_docket reads two files from a sample data
folder, preprocesses the data, converts to json format (dictionary of
dictionaries) and calls json2fp to generate fingerprint vectors.
Subsequently, the vectors are clustered and the lines of text from the
input files are printed to the console, reordered according to clusters.