Add file load option to skip rows based on comment character
Some data files may have comment lines that should be ignored. Add
a file load option (function argument) that allows specification of
a comment character. All rows in a file that start with the comment
character will be skipped. To support this functionality, refactor
file load methods to only operate on a single file at a time
(instead of allowing multiple files to be specified). Update
hello_docket.py example a needed.