Add Nextflow hello_docket example
Add three new Python scripts to ingest a file, transform the ingested
data to histograms (value occurrence counts), and calculate fingerprints
on the histograms (for rows and columns in the original data). Add an
encode_fp function in transform.py for encoding dictionary-of-dictionary
data as fingerprints. Update hello_docket.py to make use of the new
encode_fp function.

Add a Nextflow hello_docket script that calls these new functions in a
pipeline. Add a nextflow.config file and add $baseDir to the PYTHONPATH
environment variable so that Python modules can be found during Nextflow
script execution.