Add initial clustering support with Jupyter notebook for visualization
Add initial support for clustering PCA results as well as raw (original)
data. Include an example Nextflow script to execute a basic clustering
pipeline. Also, include a Jupyter notebook that gets copied to the docket
output container and allows visualization of clustering results on PC1,
PC2, and PC3 with coloring for 2, 3, and 4 clusters (both on the row-wise
and on the column-wise data).

Add additional support for using pandas for file IO, as well as better
handling of zipped and unzipped file IO. Rename load.py to file_io.py
(and refactor as necessary) to reflect more general use of this module
for both file input and output. Add a notebook folder with a template
Jupyter notebook for viewing clustering results. Add a sample data set
(in the data folder) to use with the new clustering Nextflow script
example.