Add initial code for calculating enrichment statistics
Add code to identify clusters (from hierarchical clustering) that are large
enough to get meaningful enrichment information. Also, identify attributes
for which there are sufficient numbers to calculate meaningful enrichment
statistics. For each attribute and cluster, repeatedly sample (the default
is 50 times) with replacement on the full data set (where the sample size
is the size of the cluster), calculate the fraction observed for each
attribute value, and use the sorted samples fraction values to estimate
the significance (quantile value) of the observed fraction for the given
attribute within the cluster (for each cluster).