Finish code for enrichment calculations and visualization of results
Enrichment analysis requires identification of categorical features (or
attributes) in a data set. Clustering can be done on numerical features
after which enrichment analysis can quantify statistically-significant
over- or under-representation of categorical feature values (e.g. male
or female) within the clusters. To support this type of analysis,
implement preprocessing routines to ingest one or more files (all of
which can contain a mixture of categorical and numerical features),
eliminate low-information features (e.g. based on the fraction of
missing values), and write categorical and numerical feature data to
two separate files.

Implement enrichment analysis code, including: (1) Functionality to
identify categorical attribute values for which there are sufficient
counts in the input data to make enrichment analysis (on that feature)
meaningful and (2) a resampling routine to resample with replacement on
the whole data set (using cluster size as the sample size) and, based
on that, determine the significance of the observed fraction (within
the cluster) of individuals for a given attribute value (e.g. female).

Replace the previous clustering Nextflow example with an expanded
example that does both clustering and enrichment analysis. Create a new
Jupyter notebook template that allows visualization of clustering and
enrichment results. Implement this new example such that supporting
code in the Jupyter notebook is automatically copied to the publish
directory so that functions do not need to be defined within the
Jupyter notebook.

Lastly, it was found that the Python implementation of PCA analysis
saved results in a format that was incompatible with the format
generated by docket_study. Make appropriate changes to fix this.