Add line to docket_study so format can be specified
The docket_study Nextflow script contains the following line to allow
specification of the format of the file to be ingested:

params.format = 'table'

Add the following line under /* interpret parameters */ to make sure
that format can be specified as a command line argument:

format = params.format

Note that if pandas is specified as the format, this makes it so
docket_study properly ingests comma-delimited (i.e. csv) files.